Reintroduce rpm signing at publish time.
We reverted the rpm signing during publish time
to avoid the addsign here failing due to a sign
already present(added by goreleaser when nfpm signing
works fine) - For most cases this should be okay, but
we're reintroducing this in case it fails again like before
on goreleaser's end.

We'll now check if the signature is already present before
invoking the rpm command to add signature.